+++ src/Models/Category/Category.php
@@ -4,6 +4,7 @@ namespace Expressionengine\Coilpack\Models\Category;
 
 use Expressionengine\Coilpack\FieldtypeManager;
 use Expressionengine\Coilpack\Model;
+use Expressionengine\Coilpack\Models\Channel\ChannelData;
 use Expressionengine\Coilpack\Models\Channel\ChannelEntry;
 use Expressionengine\Coilpack\Models\FieldContent;
 
@@ -46,28 +47,30 @@ class Category extends Model
         $field = $manager->getField($field, 'category');
         $column = "field_id_{$field->field_id}";
 
-        // If this field is not storing it's data on the category_field_data table we
-        // will join the separate data table with a unique orderby_field_name alias
-        if ($field->legacy_field_data === 'n' || $field->legacy_field_data === false) {
-            $alias = "orderby_{$field->field_name}";
-            $column = "$alias.$column";
-            $this->scopeJoinFieldDataTable($query, $field, $alias);
-        }
+        // Setup our join for the appropriate channel data table and alias
+        $alias = $field->hasLegacyFieldData() ? 'orderby_legacy_data' : "orderby_{$field->field_name}";
+        $this->scopeJoinFieldDataTable($query, $field, $alias);
+        $query->select($this->qualifyColumn('*'));
 
-        return $query->orderBy($column, $direction);
+        return $query->orderBy("$alias.$column", $direction);
     }
 
     public function scopeJoinFieldDataTable($query, $field, $alias = null)
     {
-        if ($field->legacy_field_data == 'y' || $field->legacy_field_data === true) {
+        // If we have already joined this table alias just do nothing
+        $alreadyJoined = collect($query->getQuery()->joins)->pluck('table')->contains(function ($joinTable) use ($alias) {
+            return strpos($joinTable, $alias) !== false;
+        });
+
+        if ($alreadyJoined) {
             return $query;
         }
 
-        $table = $field->data_table_name;
+        $table = $field->hasLegacyFieldData() ? (new ChannelData)->getTable() : $field->data_table_name;
         $joinTable = $alias ? "$table as $alias" : $table;
         $alias = $alias ?: $table;
         $query->leftJoin($joinTable, "$alias.cat_id", '=', $this->qualifyColumn('cat_id'));
-        $query->select('*', $this->qualifyColumn('cat_id'));
+        $query->select($this->qualifyColumn('*'));
 
         return $query;
     }